+++ apps/api/endoora_api/settings/test.py
@@ -1,4 +1,14 @@
+import os
+
 from .base import *  # noqa: F403
 
 DEBUG = False
 PASSWORD_HASHERS = ["django.contrib.auth.hashers.MD5PasswordHasher"]
+
+if os.getenv("ENDOORA_TEST_DATABASE", "sqlite") == "sqlite":
+    DATABASES = {
+        "default": {
+            "ENGINE": "django.db.backends.sqlite3",
+            "NAME": ":memory:",
+        }
+    }